+++ python-libraries/data-platform-catalogue/data_platform_catalogue/client/datahub/search.py
@@ -2,9 +2,10 @@ import json
 import logging
 from datetime import datetime
 from importlib.resources import files
-from typing import Any, Sequence
+from typing import Any, Literal, Sequence
 
 from data_platform_catalogue.search_types import (
+    FacetOption,
     MultiSelectFilter,
     ResultType,
     SearchResponse,
@@ -62,6 +63,7 @@ class SearchClient:
 
         page_results = []
         response = response["searchAcrossEntities"]
+        facets = self._parse_facets(response.get("facets", []))
 
         logger.debug(json.dumps(response, indent=2))
 
@@ -80,7 +82,7 @@ class SearchClient:
                 raise ValueError(f"Unexpected entity type: {entity_type}")
 
         return SearchResponse(
-            total_results=response["total"], page_results=page_results
+            total_results=response["total"], page_results=page_results, facets=facets
         )
 
     def _map_result_types(self, result_types: Sequence[ResultType]):
@@ -197,3 +199,31 @@ class SearchClient:
             tags=tags,
             last_updated=last_updated,
         )
+
+    def _parse_facets(
+        self, facets: list[dict[str, Any]]
+    ) -> dict[
+        Literal["domains", "tags", "customProperties", "glossaryTerms"],
+        list[FacetOption],
+    ]:
+        """
+        Parse the facets and aggregate information from the query results
+        """
+        results = {}
+        for facet in facets:
+            field = facet["field"]
+            if field not in ("domains", "tags", "customProperties", "glossaryTerms"):
+                continue
+
+            options = []
+            for aggregate in facet["aggregations"]:
+                value = aggregate["value"]
+                count = aggregate["count"]
+                entity = aggregate.get("entity") or {}
+                properties = entity.get("properties") or {}
+                label = properties.get("name", value)
+                options.append(FacetOption(value=value, label=label, count=count))
+
+            results[field] = options
+
+        return results